requirements.txt

fastapi==0.112.2
uvicorn==0.30.6
prometheus-client==0.20.0
pydantic==2.8.2
structlog==24.4.0
google-generativeai==0.8.2
python-multipart==0.0.9
streamlit==1.37.1
